Use wxString::StartsWith()
Replace the pattern
	if (myString.Mid(0, 6) == wxT("foobar")) { ...
with
	if (myString.StartsWith(wxT("foobar"))) { ...
This more clearly expresses the intention of the code. It's also less fragile
because it doesn't require the length parameter in Mid() and the actual length
of the string literal to be in sync. And as a bonus it's also more efficient
because it doesn't require creating/destroying a temporary substring.