Move most functionality from CatapultPage to openMSXController
CatapultPage is an (abstract) base class for AudioControlPage, SessionPage,
VideoControlPage, ... Before this patch the main thing it did was respond to
setting-update-events from openMSX (so update the GUI when some setting has
been changed in openMSX itself). To do this (much simplified) it has a list of
setting names with associated callback functions that each update some
particular GUI widget. Though this list is duplicated in each CatapultPage
object. Also when the openMSXController receives an incoming setting-update,
it sends it to just any of the CatapultPage objects, currently it picks
VideoControlPage. So CatapultPage handles both video and audio (and other)
settings. When it turns out it's an audio related setting, it still has to
'lookup' the AudioControlPage (and this requires those weird assert-that-
audiopage-is-not-null checks).

Anyway, this all indicates that this functionality was in the wrong location.
I've moved it to openMSXController (this was the class that originally passed
the event to CatapultPage). So now the callback information is no longer
duplicated. Also openMSXController has (indirectly) access to e.g. audiopage,
so it's no longer needed to look it up by name (and assert that it's found).

This patch also splits updateAudio() into updateVolume() and updateBalance().
Now that this audiopage-lookup is no longer needed, the only thing
updateAudio() did was figure out whether it should do volume or balance stuff
and then handle that. We can as well assign volume and balance settings to
different callbacks.

While debugging some problem I had while developing this patch, I also
simplified the openMSXController::translate() method. This is part of the
(very ugly/hackish) launch script stuff. It recognizes commands that start
with 'magic-markers' and does 'special' things on them (this should all be
rewritten in the future). I've simplified this to only detect those
magic-markers in the first token in the command (because at some point it
recognized the '!=' (not-equal) operator as a special command).