wxString::Find() related cleanups
Before this patch, the code mostly tested for '-1' to check if a value was not
found. Though the wx documentation says you should use wxNOT_FOUND. Obviously
in the current version of wxWidgets 'wxNOT_FOUND == -1', but that might change.

In the openMSX code we already had problems with this in the past. For
std::string::find() we also tested on '-1' instead of 'string::npos'. But when
switching from 32 to 64-bit platfrom, string::npos is no longer the same as
int(-1). Of course also the type of the return value has to change (from int
to string::size_type), wxString::Find() does not use any typedefs.

I also replaced
	string.Find(..) != wxNOT_FOUND
with
	string.Contains(...)