Merged openMSXController, openMSXLinuxController and openMSXWindowsController
Before this patch openMSXController was an abstract base class with two
concrete implementation (linux and windows). Though any particular build will
only either have the linux subclass or the windows subclass. We never need
both in the same application. So we also don't need this class hierarchy with
associated virtual functions.

So this patch merges the implementation of openMSXLinuxController and
openMSXWindowsController into the openMSXController base class and removes the
class hierarchy. To distinguish between linux/windows there are now some more
ifdefs in the code. Because the linux/windows code is now closer together it's
easier to see (and factor out) common code between both versions.

This patch (obviously) touches windows specific code. I can't test this, so I
hope I didn't break anything.

--

This patch also passes the openMSXController object by reference instead of by
pointer. Pure technically a pointer and a reference are the same (under the
hood a reference is implemented as a pointer), but semantically there are some
differences:
 - A reference cannot be null. So it's always safe to use a reference, while
   for a pointer you always have to think about "can I dereference this or
   should I first check for nullptr".
 - Once initialized a reference cannot be 'reassigned' to 'point to' another
   object. So if a reference is initialized at the top of a (very long) function
   you can be sure that that same reference still refers to the same object at
   the end of that function (without having to examine all code in between).
Both these restrictions make the code easier to reason about when using
references.