Added some comments
Vampier did all the work, I only want to add some cooments for future
reference.

After revision a91edd the windows compilation errors were solved, but it still
wouldn't link. Eventually Vampier found the solution:
  http://forums.wxwidgets.org/viewtopic.php?t=34044&p=140263
It seems that the '#include <windows.h>' declares some macro that clash with
names that are also used inside wxWidgets headers. This causes the class
definitions to look different between classes that (indirectly) did or did not
include windows.h. And that in turn gives linking errors because the ODR rule
is violated. Anyway, seems to be a common problem because wxWidgets ships a
special header file that undoes the damage caused by windows.h.