Avoid visual studio compiler warning about strcpy/strcat
Visual stdio complained about the use of strcpy() and strcpy() (something about
security and deprecation). In general you indeed have to be careful with buffer
overflows when using these functions. Though in this case the code was OK.

I still rewrote this code. Because it's good to have a warning-free build
(but I believe there still are more visual studio warnings) but also because
this could be written more efficiently:
- Avoid heap allocation/deallocation, we now allocate on the stack.
- Avoid calling strlen() twice, the result of the 2nd call is known.