Cleanup wxArrayString related things
In modern c++11 when using STL containers, the (new) preferred style is to
return such containers by value instead of using an out-parameter. The reason
is that this is a much more natural syntax. And since C++11 STL containers are
now move-enabled, this is just as efficient.

The wxArrayString class is NOT move enabled (in future versions of wxWidgets it
likely will be). So we do pay the (small) cost of extra copies, but for the
purpose of wxCatapult, code clarity is more important than the (small)
performance overhead (the containers are usually very small).

--

Some getter methods only returned (a copy of) some private wxArrayString member
variable. In such cases it's better to instead return a const-reference to this
variable. (And for the caller this has the same natural syntax as in the
previous section.)