Moved CatapultPage::ConvertPath() to utils
There was already a TODO about this in the code.

This function had 2 boolean parameters (each with a default value). The
function did very different things depending on the value of these booleans.
The functions was also always called with compile-time-constant boolean values
(I mean either 'true' or 'false',  but never a value from a variable). (Also
only 3 of the 4 possible combinations were used). So I split this function in 3
separate functions and removed the booleans.

There is a new file utils.cpp. I hope I correctly added this to the msvc
project file.