Use proper Tcl-list-parser to handle result of launch script
The launch script performs several queries on openMSX via Tcl commands. Some of
these queries return a list of values. Such lists are formatted as Tcl lists
(normally the items in the lists are separated by spaces, but if the items
themselves contain spaces there are a few quoting/escaping rules). Before this
patch we avoided having to parse Tcl lists: instead of a direct query, we send
a command that also split the result on newline characters, so each item of the
list now appears on a separate line (as long as the items themselves don't
contain embedded newline characters this is a reasonable approach). Though just
doing the proper Tcl parsing is in the end a better solution, so that's what
this patch does.

The current launch script implementation is quite ugly. One aspect of this was
that all commands that start with '#', '^' or '!' were transformed (to do this
newline stuff). This transformation aspect is now removed. Another ugly hack
was that the result of such commands is handled special. The result is stored
and used in later commands that contain special '*' characters. I'll remove
this '*'-hack in later patch(es), for now I only simplified how these results
are detected and stored. Instead of recognizing special start characters, the
(few) commands that actually need to store the result now do this manually
in their callback function. Also instead of storing the raw returned string and
re-parsing it each time it's used, we now store the parsed result.