Simplify executeLauch() / return void from launch callbacks
ExecuteLauch() did two very different things depending on whether this method
was called with a nullptr or not. At the end of this function there was also a
common part. I've now split this method in 3:
- ExecuteStart(): when old version was called with nullptr
- ExecuteLauch(): when old version was called with non-nullptr
- ExecuteNext():  the old common tail part

All the lauch script command callback functions currently return the value '0'.
This value indicates how many instructions in the script to skip. (Older
revisions did sometimes return non-zero values). Because this mechanism is
not longer used I've removed it. It was also very fragile.