Removed 'showerror' from lauchscript
The largest part of this patch changes the layout of the calls to
AddLauchInstruction(). Now it's easier to see what kind of values are passed
to each of the parameters of this function.

One of the parameters was 'showError'. If a command in the lauch script
generates an error, this parameter determines whether this error can be ignored
or whether it should be printed. Though each time such an error would be
printed it's actually a bug in wxcatapult (so this is a kind of debug feature).
Very few commands actually used this 'showError' flag (even commands that
should never fail). Because this feature isn't used very much and because I
really prefer to simplify this lauch script as much as possible, I decided
to remove it.