Misc launch script: lambdas / parameter / info exists
Use std::function instead of member-function-pointers to store the callback
functions in launch script commands. This is a new c++11 feature. (It works
with gcc/clang and I think it also already works with msvc). Next to member
functions, std::function allows to pass any callable entity with the same
signature. More specifically this allows to use lambda functions (another new
c++11 feature). This is useful for several reasons, see below.

The AddLaunchInstruction() function took a parameter with the name 'parameter',
this value was later passed (possibly with some transformations applied) to the
callback. But now that we have lambda functions we can instead write the
required parameter directly in the lambda. Many launch commands didn't actually
use this 'parameter', so not requiring it is a simplification.

The launch script mechanism recognized the 'info exists' command and handled it
special: if this command returned '0' it was treated as-if this command
returned an error and this in turn (in this specific case) caused the next
command in the launch script to be skipped. This is again a very ugly and very
fragile hack. And it wasn't even necessary: the typical use case was:
  info exists renshaturbo    -> in callback enable the GUI widget
  set renshaturbo            -> set current value in GUI widget
so the second command was skipped if the setting doesn't exist. If we don't
skip this command then that command fill fail, but that doesn't matter, such
errors are ignored anyway (the launch script already uses this ignore-errors
approach in other cases). So this patch completely removes the special casing
for 'info exists'.

+ simplify 'info exists'