Removed @checkfor
The launch script mechanism had another hack to handle '@checkfor' special.
This was not a command that was send to openMSX, instead it conditionally skips
the next instruction(s). Though just like in the previous patch it's simpler
to always execute the instruction and ignore the possible error.

@checkfor also used the 'lastdata' internal state. I'd like to completely
remove this hack (eventually), so this patch is a step in the right direction.