Remove @execute, instead issue empty command
After all the previous simplifications, the only special command in the launch
script that started with a '@' character was '@execute'. This command was not
send to openMSX, instead it was used to trigger a c++ callback when a certain
point was reached in the launch script.

There was special logic to detect this special command, skip over it, fake a
success return value, etc. It all becomes much simpler if instead we send a
real but empty command to openMSX. This may be slightly less efficient because
we introduce another catapult-openmsx-catapult round-trip, but this overhead is
hardly measurable and in the long run, I believe, all these former '@execute'
commands should go away completely.

This patch completely removes the special meaning of the '@' character from the
launch stuff.