Removed action field from launch commands
After all previous simplifications the 'action' field of each launch command
was only used (anymore) to skip the next 'n' instructions in case a command
triggered a (Tcl) error. (And this value 'n' was actually wrong in some cases).
Though just like in the previous patches, instead of trying to skip commands
it's much easier to always execute the commands and ignore (potential) errors.