Make better use of lambda functions
One of the previous patches replaces member-function-pointers with
std::function and lambda functions. Though basically the only thing the lambda
functions did was call the original member function. Very often that member
function was only one or a few lines of c++ code. This patch will (where
appropriate) move that member function directly into the lambda function.
Because the launch Tcl command and the corresponding c++ action are now much
closer together the code is a lot easier to follow.