Also use std::function/lambda for settings
Next to the launch script openMSXController has a similare list of setting
names and corresponding actions it should perform when a openMSX setting
changes. Before these actions were implemented as member-function-pointers,
this patch changes it to std::function + lambda functions (just like for the
launch script commands).

Just like for the commands, this change allows to:
- Remove parameters from the AddSetting() function (the 'control' and 'flags'
  parameteres are gone).
- Simplify the callbacks, e.g. only pass parameters that are actually used, or
  inline the functions that are only used once directly into the lambda.