Use callbacks in WriteCommand / simplify update handling
The previous commit explained there were two main users of WriteCommand(). The
routine that handled the replies for non-launch-commands had an ugly list of
hard coded command strings. This patch replaces that hard coded list with
callbacks that are passed by the original caller of WriteCommand(). This makes
the code much easier to understand (only need to read very local code instead
of code spread over different source files).

This patch also simplified the openmsx-update-event handling code. This code
sometimes has to check that e.g. a setting update is not triggered by the
command that was just send to openMSX (to avoid infinite loops or strange delay
effects). To do this the old code did various string manipulations on the full
command string. But it's much easier to first parse the full command into
tokens and then check specific tokens (just like how a real Tcl interpreter
analyzes a command). This parsing into tokens is exactly the same as parsing a
Tcl list into elements, and we already have a routine to do this.