Remove loop handling from launch script code
The previous patch already explained that '*' was still handled special in a
launch command. When this character was seen, it got replaced with each element
from a previously returned list. This patch completely removes that special
logic from the launch script code and instead implements it in specific
callback routines. Implementing it there was actually fairly easy now that we
have an improved WriteCommand() function with callbacks. And as a bonus it's
now trivial to re-introduce the 'parallelism' that was removed in the previous
patch.

This removed the last 'special interpretation' that the launch script code gave
to the commands. Now the launch script is just a collection of Tcl commands that
are send (unmodified) to openMSX. Each command has an associated callback that
is executed on a (successful) reply on the command. The next command is only
send after a reply of the previous command is received (successful or not).