Simplify GetOpenMSXVersionInfo()
The GetOpenMSXVersionInfo() method had a windows and non-windows specific
implementation. Though the windows version only used generic wxWidget functions,
so it can work as well on non-windows systems.

So this patch removes the non-windows specific (actually non-generic) code.
Because GetOpenMSXVersionInfo() is now so small I've inlined it in the (only)
caller.

An extra advantage of the generic version is that it doesn't leave a temporary
file '/tmp/catapult.tmp' behind (the code used this file but forgot to delete
it).