Split StartOpenMSX() in two
This function took a boolean (2nd) parameter, depending on this value the
function did two very different things. Basically the whole function was a
if-then-else statement with one function in the then- and the other in the
else-block. There was only a single common statement in front of this if, but I
believe that statement only makes sense for one of the two blocks. (Ok, there
was also some boilerplate code around the if to handle the return value, but
actually only one of the two functions needed a return value).

So now there is a new function CheckVersion(). This will start openmsx with the
option '-v' (this will only print version information and exit). The function
then parses the output and checks for a minimum openMSX version. That's it. The
old function also executed some setup code (e.g. enable some GUI button), that
code is now moved. I think this was actually a bug when this function was
called via the 'Setting > Edit Configuration ...' menu.

The other part is still called StartOpenMSX(). This will actually launch
openMSX (and leave it running) and setup the communication channel(s) with the
openMSX process.