Cleanup openMSXController <-> wxCatapultFrame interaction
The class 'openMSXController' should concentrate on the interaction with the
openMSX process, the class 'wxCatapultFrame' should concentrate on the GUI
stuff. Though in the current code these two tasks are not well separated.

This patch adds a cleaner separation during start/stop. This allowed to replace
a bunch of public methods with only two new public methods: OpenMSXStarted()
and OpenMSXStopped(). The former public methods are now private. There was even
a public member variable, that's now also private.