Factor out common code bewteen {disk,cart}{a,b}
The code that handled diska was copy/pasted to handle diskb. This patch tries
to have only one set of routines that are parameterized on the disk (a or b).
Similar for carta and cartb.

There still is code duplication between disks, carts and cassette. E.g. the
code to insert/remove a disk/cart/cassette is still very similar (but not
completely the same).