Various simplifications in WriteMessage error handling
There was a very complex routine to transform the error code 'errno' into a
wxString message. It used 'strerror_r()' and had logic to deal with dynamically
growing buffer sizes. It also dealt with the difference between the XSI and the
GNU variant of this function. But there is a much simpler function: strerror().
This function is not thread-safe, though as long as you only access it from the
main thread (as is the case for wxcatapult) then everything should be just
fine. (BTW: the buffer management in the old code was also rather bad. In
modern c++ code (especially c++11), you should rarely have to write 'new' and
'delete' explicitly, instead you should use 'managed' (memory) resources. In
this specific case 'vector<char>' could be used to manage the memory
allocation.)

Never use exception specifications (throw annotations) in c++ code. It was a
bad idea in the c++ standard. It's a deprecated feature in c++11. This (very
old) article explains why:
   http://www.gotw.ca/publications/mill22.htm
Very short summary: c++ and Java exception specification syntactically look
similar, but it are very different things:
  Java: they are statically checked by the compiler
  C++: they are dynamically checked at run-time
When in Java you violate an exception specification, the program won't compile,
in c++ the code compiles (even without warnings) but at-run-time the program
gets terminated.

If you inherit from std::exception (or any of the more specific exception
types), you're expected to override the what() method. This returns a 'const
char*' with a description of the error that occurred. Though the standard
doesn't say anything about the encoding of this string. Also we want this
message as a wxString and it would be weird to convert the message back and
forth between 'const char*' and 'wxString'. So I just dropped the inheritance
from std::exception. Since we don't have a top-level catch block that attempts
to handle all possible exceptions, there was anyway no advantage in using
std::exception.

In c++ it's not needed to write empty constructors/destructors. It's also not
needed to explicitly call the default constructor of a base class. Especially
in very small classes (like the exception classes in this patch), the general
practice is to leave out these redundant constructs.

See also next patch for more on the exceptions...