Don't use exceptions to handle WriteMessage() errors
We briefly discussed this on IRC, and I agreed that using exceptions for this
seemed like a good idea. Though now that I've seen the actual implementation I
have to change my mind:

Exceptions become useful when the throw-site and the catch-site are far apart
in the code (lot's of function calls in between). Or when the caller has more
specific information it can add to the error. Or when the error has to be
handled in a different way depending on the context. In the current code none
of these reasons apply: WriteMessage() throws an exception. There are only two
direct callers and both immediately handle the exception in exactly the same
way.

The code becomes much simpler if instead of throwing an exception (and already
one level up handling it), we directly handle the error in WriteMessage().
(This also makes it easier to see that the actual printed message was quite
redundant.)

Also the exception class inheritance hierarchy was (IMHO) over-designed: there
was an abstract base class, one subclass that holds a string and another that
holds an error code. If there are lots of exceptions (stored in memory at the
same time) or if we try to catch very specific exceptions (bad idea in
general), then this design makes sense. Though in this case the only thing we
ever did with the errno-exception was turn it into a string. So we could as
well do that while constructing the exception, and then only have one exception
class (without inheritance) that stores this string.

So together with the previous patch this implements the error handling with
only 15 lines of extra code ;-)  If in the future we refactor the code so that
WriteMessage() and the location where the error is handled are further apart,
then (part of) this patch may be reverted. But I still think the inheritance
thing was overkill. (E.g. we might want to kill the openMSX process and revert
the GUI to the 'stopped' state in case of an error .. although I think the case
were the openMSX process dies is already handled fine, so I prefer to first see
an example of another scenario where this error can occur before complicating
the code.)