Hack: don't delete m_connectThread
Vampier reported a crash on exit. This was in windows-specific code while
destroying the m_connectThread object. I recently added this because before it
was leaked. Though a leak is better than a crash, so I restored the old
behavior. It's not a big leak: only a few kB (or less?) each time openMSX is
stopped. Someone with a deeper understanding of wxWidgets could look into this,
but for me this is low priority.