Simplify 'type ;' fix
The intention of the tclEscapeWord() functions is to escape _all_ characters
with special meaning in Tcl. So that just _any_ string can be passed as a
(single) parameter to a Tcl proc. Including strings that contain ';' characters
(the ';' character separates two Tcl commands (just like '\n')).

So the previous patch did fix the type command, but it did not fix all other
commands that used tclEscapeWord() in combination with ';'.