Do not attempt to delete a detached wxThread
A detached thread deletes itself when it completes. Therefore it is
not safe to delete it from another thread. And it is certainly not
safe to delete it directly from the heap rather than through
wxThread::Delete.
  http://docs.wxwidgets.org/trunk/classwx_thread.html#thread_deletion

Instead of keeping a reference to the thread around, I just fire it
and forget about it. If the thread takes a long time to connect and
another connect thread is created before the first one finishes, there
could be problems, but I don't think the old code handled that
gracefully either.