Make sure pipe is created before process is spawned
I took the CreateNamedPipe() call out of CreateControlParameter() since
it didn't belong there; the inability to abort on error showed that.
However, I moved it a bit too late in the startup sequence: the named
pipe should be created before CreateProcess() is called, since the
spawned process will want to connect to it.