Fixed crash on switching broken->working machine
(As already mentioned in previous patch) steps to reproduce:
 - start/stop a working machine
 - start a broken machine (will not work, so stops automatically)
 - restart the working machine
    -> catapult crash

I bisected the problem to revision f6df2f76b5c092a9381d55d5375243c782305c52.
That patch introduces a 'm_launchCallback' queue. The bug was that on
HandleEndProcess() (triggered early by the broken machine) we didn't clear
this queue.

As mentioned in f6df2f76, it would be much cleaner to not have m_launchCallback
and instead use the 'normal' command queuing (that would have avoided this
bug).