Restored revision extraction.
It was commented out previously, because it didn't work on MinGW32.
However, I tried to implement a solution for that. Because I do not have
such an environment, I couldn't test that solution. Once someone starts
with MinGW32 development of Catapult again, this should be verified. At
least it works again now for the other platforms.