Fix text shown in disk combo box
Problem was reported here:
	http://www.msx.org/forum/msx-talk/openmsx/latest-catapult-version#comment-228405
To reproduce insert a disk via catapult twice. The disk is correctly inserted
on the openMSX side, but the combobox shows a blank string instead of the
path to the disk image file.

This regression is caused by revision ffab1d93. Part of that patch disables
reacting to openmsx update events when that event was triggered by a command that
was send by catapult itself. (Unfortunately the commit message doesn't explain
exactly _why_ this is required).

Though the heuristic to detect that the 'diska media update' was triggered by
the catapult diska-browse-action was broken. When you insert a disk via the
catapult disk browser it will send 2 commands to openmsx:
	diska eject
	diska your-selected-disk-image.dsk
and openmsx will reply with two update messages:
	<update ... name="diska"></update>
	<update ... name="diska">your-selected-disk-image.dsk</update>
On both update events the heuristic kicks in. In both cases the last command
executed by catapult is "diska your-selected-disk-image.dsk". So on the first
update event the heuristic concludes the update and the last command are
unrelated, so the event is processed and the combo box is cleared. On the
second event the heuristic concludes this update is caused by the last command
and the event is not processed further. So the combo box remains blank.

I've fixed this by not executing the (redundant) 'diska eject' command. And
then the heuristic works fine. It might be better to remove this heuristic
completely. so just process all update events. Unfortunately I do remember this
caused problems (for (some?) settings?), but I don't remember what exactly :(