Refactor cassette widget enable code
Move _all_ logic that controls enable/disable state of cassette-related
widgets to one central SetCassetteControl() method. Make sure this method
gets called on all conditions that influence the enable/disable state.

Reimplement the way in which we detect whether the MSX machine has a
cassette player (and store that in a boolean instead of a string).

Don't try to auto-plug the cassetteplayer in OnBrowseCassette(), just do
it once when openMSX starts. (Why did the old code do it in OnBrowseCassette()
but not in any other code path that inserts a cassette file?)