Always call OnChangeCassetteContents() from OnClearCassette()
According to the comments this is only needed on windows. Even if this is
(still) true (I doubt it is) it's a very good idea to keep the code as much as
possible the same on all systems.

I actually think it _is_ required (now) to call OnChangeCassetteContents() (and
indirectly SetCassetteControl()) to get the button-enable status correct (the
state of some buttons depends on the cassette filename being empty).