Also use insertMedia() in Click{Disk,Cart,Cassette}Combo
Remove code duplication from even more locations.

There is one functional change though. Before when you selected a media from
the history, the history was left unchanged. Now the last selected is moved
to the top of the history (so history is sorted on least-recently-used), but
that may actually be a good thing.