Clear IPS status in insertMedia()
All callers of insertMedia() already did this (clear ips), except for
checkLooseFocus(), but I believe that one should clear it as well. To make this
possible we need to know the name of the 'ips label', so added that to the
MediaInfo struct.

Factor out code to set the mapper type. Renamed this routine from
UpdateMenuMapperLabel() to SetMapperType(). For this added 'type label' to
the MediaInfo struct.

There was a bug in ClickCartCombo(), the mapper type stored in history was not
correctly restored.