Run-time insertion of cart{a,b} didn't pass romtype
It only worked for ROMs that are already inserted when openMSX starts. Luckily
it's easy to fix now that the code is in a central insertMedia() routine.

When the mapper-type is changed on an inserted ROM (so when openMSX is running)
we should send a new command to openMSX.

Similarly IPS patches are not applied to ROMs that are inserted at-run-time
(only for ROMs present when openMSX starts). I'll fix that in a later patch.