Remove MediaType::contents member variable
This string is always in sync with control.GetValue(). So no need to store
both.

This simplifies (removes) the code that should keep the old 'contents' variable
in sync. But it makes the code that used that variable slightly longer:
  m.contents
vs
  m.control.GetValue()

By making this change I discovered a bug: when changing the list of IPS patches
(on disks) while openMSX is running, the command we send to openMSX didn't
separate the IPS filenames by spaces (so more than one patch wouldn't work).