Remove 'wxArrayString history' member variable.
Each 'media' kept a list of recently inserted filenames both in the combobox
widget and in a wxArrayString (exactly the same string). So remove the
redundant wxArrayString. And simplify the code that kept both in sync.

Also simplify the implementation of AddHistory(): apparently early versions
of wxWindows didn't support inserting in a wxComboBox yet.