Remove all OnChangeXXXContents() methods
These methods are used to clear the IPS list and mapper type when typing a new
filename. I think it's fine if we only do this once the user finishes typing
the whole filename. And InsertMedia() already takes care of this IPS-clearing,
so we don't need anything special anymore.

The extra advantage is that when you start typing, then realize you're mistaken
and restore the filename. Then you don't loose the IPS list and mapper type.