Unify OnBrowse{cart,disk}Ips / refactor insertMedia()
There was an OnSelectIPS() method that handled IPS patch selection for
cartridges and an OnBrowseIps() method that did the same for disks. Only the
latter would re-send the insert command to openMSX when the list of IPS files
changed (that became a bug once we added support for run-time insertion of
cartridges). This patch unifies both into one new OnBrowseIps() method. And
run-time IPS selection for cartridges should also work now.

OnBrowseIps() is another method that sends an insert-media command to openMSX.
This patch refactors insertMedia() so that it can be reused in this new context
(e.g. it now knows about the -ips command flag).

The old insertMedia() method is split in two:
  insertMediaClear() and insertMedia()
The former clears the IPS list and mapper-type and then calls the latter.

The old insertMedia() method would clear the IPS list but not the mapper-type.
That's actually strange: we should either clear both or neither. The old code
did this, always clear both, except in one situation: when we select a
cartridge from history, we restore the mapper-type but clear the IPS list. This
is actually bug: we should also restore the IPS list from history, but that's
something for a later patch.