Various cleanups related to string splitting and iterating
Catapult has some ad-hoc way of encoding a collection of strings (filenames,
mapper-types, ...) into a single string: concat the elements separated by a
double colon. So something like:
  elem1::elem2::elem3::
This means it's not possible to have items that contain themselves a double
colon. Luckily this doesn't happen often in filenames. But it does make it
impossible to have nested structures (in the near future we'll need a
list-of-list-of-ips-files). There are many other encoding that don't suffer
from this problem (one example is a Tcl list). These strings are stored in
configurations files, so it's not easy to change this encoding without breaking
backwards compatibility.

Anyway, enough ranting. This patch adds a helper function split() that decodes
such a string back to individual elements. (Because such decoding logical was
duplicated in quite a few places).

This patch also simplfies some for-loops: it's possible to use c++11 range
based for loops also on wxwidget-collections.