Removed UpdateSessionData() and related member vars
The UpdateSessionData() method copied some state from the GUI controls to the
member variables m_InsertedMedia, m_usedMachine, m_usedExtensions. These
variables were later only used to store in the configuration file. Though there
was a bug that didn't call UpdateSessionData() often enough(?), and the saved
configuration didn't reflect the most recent catapult state.

However there's really no need to have a UpdateSessionData() method and extra
member variables. We can instead extract the information just before storing
it to the config file. Then those variables can never be out-of-date.

There's one other change in this patch: move the SaveHistory() call from
AddHistory() to insertMedia(). AddHistory() was (only) called from
insertMedia() in case the media was not empty (not ejected). But the saved
history also contains a bitvector with the inserted media, that one should
also be updated when the media is ejected.