Don't plug cassetteplayer from catapult
When catapult detects openmsx has just booted, it send the command
   plug cassetteport cassetteplayer
Not sure why .. maybe this is still from the time we didn't have an
autoplug Tcl script yet.

So this command isn't needed anymore. But it also interrupts a replay when
you reverse all the way to the start (because then catapult detects a freshly
booted MSX and it resends the command).