Added super ugly hack to workaround extensions reported broken
Some extensions only work on certain machines and typically it's not the
machine that Catapult is using to test the extension with. However, the
goal of the testing was almost exclusively to find out about
missing/uninstalled system ROMs. As these extensions do not have
system ROMs, we treat them specially: we just say they are working
TODO: implement a better solution.