Get machine/extension list directly from openMSX
This also eliminates the share dir setting, which was confusing our
users, as they thought they could use it to configure openMSX, while it
was only meant to let Catapult find the list of extensions and machines.

Catapult now does this by asking openMSX the list via the -bash command
line option.