Do not unplug joyporta/b/printerport for nothing.
People noticed that when they loaded a replay after having openMSX started from
Catapult, it was immediately interrupted. This was caused by the InitJoystick
method which for some reason always unplugs when the current value of the
combobox is --empty--... It should only do that when the value changes, not
initially.

Actually, this problem still occurs if you select something in Catapult for
joystick or printerport and then load the replay from the console. Catapult
will plug in the pluggable and you'll get an interrupted replay... I didn't
feel like fixing this now, as it's far less likely to happen. (Note that a
workaround is anyway to use the -viewonly option.)

Also, the method name was confusing, it was also used for the PrinterPort.
Renamed accordingly.